Making new folders and refactoring
