+++ lib/widgets/main_widgets/base_class.dart
@@ -3,13 +3,13 @@ import 'package:provider/provider.dart';
 import 'package:universal_organizer/widgets/moving_container/settings.dart';
 
 import 'top_bar.dart';
-import 'moving_container/moving_container.dart';
-import 'moving_container/nav_rail.dart';
+import '../moving_container/moving_container.dart';
+import '../moving_container/nav_rail.dart';
 import 'google_calendar.dart';
 import 'notes.dart';
 
-import '../providers/moving_provider/nav_provider.dart';
-import '../providers/moving_provider/settings_provider.dart';
+import '../../providers/moving_provider/nav_provider.dart';
+import '../../providers/moving_provider/settings_provider.dart';
 
 class BaseClass extends StatefulWidget {
   const BaseClass({
